- Добавлено создание групп пользователей по умолчанию - Небольшие изменения стиля панели администратора - Исправлен .gitignore
