+++ apps/app_settings/apps.py
@@ -2,6 +2,20 @@ from django.apps import AppConfig
 from django.db.models.signals import post_migrate
 
 
+class AppSettingsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.app_settings"
+    verbose_name = "App settings"
+
+    def ready(self):
+        post_migrate.connect(
+            register_task,
+            sender=self,
+            weak=False,
+            dispatch_uid="app_settings.register_task",
+        )
+
+
 def register_task(*args, **kwargs) -> None:
     """Registers periodic/background tasks after migrations."""
     from django_celery_beat.models import CrontabSchedule, PeriodicTask
@@ -25,17 +39,3 @@ def register_task(*args, **kwargs) -> None:
             "enabled": True,
         },
     )
-
-
-class AppSettingsConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.app_settings"
-    verbose_name = "App settings"
-
-    def ready(self):
-        post_migrate.connect(
-            register_task,
-            sender=self,
-            weak=False,
-            dispatch_uid="app_settings.register_task",
-        )
