+++ apps/discovery/new_permissions.py
@@ -1,5 +1,5 @@
 from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Permission
+from django.contrib.auth.models import Group, Permission
 from django.contrib.contenttypes.models import ContentType
 
 
@@ -11,4 +11,11 @@ def create_permission(sender, **kwargs) -> None:
         codename="access_discovery",
         name="Can use auto discovery",
         content_type=content_type,
+    )
+
+
+def create_groups_with_permissions(sender, **kwargs):
+    g, _ = Group.objects.get_or_create(name="Доступ к обнаружению оборудования")
+    g.permissions.add(
+        Permission.objects.get(codename="access_discovery"),
     )
